Refactor: move PatternInfo to results package
