+++ src/test/java/com/suse/salt/netapi/calls/modules/PkgTest.java
@@ -1,12 +1,15 @@
 package com.suse.salt.netapi.calls.modules;
 
+import static java.util.stream.Collectors.toList;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 import com.google.gson.reflect.TypeToken;
 import com.suse.salt.netapi.calls.modules.Pkg.Info;
 import com.suse.salt.netapi.parser.JsonParser;
 import com.suse.salt.netapi.results.Change;
+import com.suse.salt.netapi.results.PatternInfo;
 import com.suse.salt.netapi.utils.Xor;
 
 import org.junit.Test;
@@ -19,9 +22,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Stream;
-
-import static java.util.stream.Collectors.toList;
-import static org.junit.Assert.assertTrue;
 
 /**
  * Pkg unit tests.
@@ -165,12 +165,12 @@ public class PkgTest {
 
     @Test
     public void testListPatterns() {
-        TypeToken<Optional<Map<String, Pkg.PatternInfo>>> type = Pkg.listPatterns(false).getReturnType();
+        TypeToken<Optional<Map<String, PatternInfo>>> type = Pkg.listPatterns(false).getReturnType();
         InputStream is = this.getClass()
                 .getResourceAsStream("/modules/pkg/list_patterns.json");
-        JsonParser<Optional<Map<String, Pkg.PatternInfo>>> parser =
+        JsonParser<Optional<Map<String, PatternInfo>>> parser =
                 new JsonParser<>(type);
-        Optional<Map<String, Pkg.PatternInfo>> parsed = parser.parse(is);
+        Optional<Map<String, PatternInfo>> parsed = parser.parse(is);
         parsed.ifPresent(pattern -> {
             assertTrue(pattern.get("base").isInstalled());
             assertFalse(pattern.get("32bit").isInstalled());
